+++ tests/conftest.py
@@ -1,8 +1,13 @@
 import importlib
 import itertools
 import json
+import logging
 import os
+import pathlib
+import sqlite3
 import sys
+import tempfile
+from functools import partial
 from unittest import mock
 
 import pytest
@@ -13,10 +18,55 @@ from nzshm_common.location.location import LOCATIONS_BY_ID
 # from pynamodb.attributes import UnicodeAttribute
 from pynamodb.models import Model
 
+import toshi_hazard_store.config
+import toshi_hazard_store.db_adapter.sqlite.sqlite_adapter
+import toshi_hazard_store.model.caching.cache_store
 from toshi_hazard_store import model
 from toshi_hazard_store.db_adapter import ensure_class_bases_begin_with
 from toshi_hazard_store.db_adapter.sqlite import SqliteAdapter
+from toshi_hazard_store.db_adapter.sqlite.sqlite_store import safe_table_name
 from toshi_hazard_store.model import openquake_models
+
+log = logging.getLogger(__name__)
+
+cache_folder = tempfile.TemporaryDirectory()
+adapter_folder = tempfile.TemporaryDirectory()
+
+
+@pytest.fixture(autouse=True)
+def default_session_fixture(request, monkeypatch):
+    """
+    :type request: _pytest.python.SubRequest
+    :return:
+    """
+    log.info("Patching storage configuration")
+
+    def temporary_cache_connection(model_class, folder):
+        log.info(f"TEMP CONNECTION for {model_class} at {pathlib.Path(str(folder.name))}")
+        return sqlite3.connect(pathlib.Path(str(folder.name), "CACHE"))
+
+    def temporary_adapter_connection(model_class, folder):
+        dbpath = pathlib.Path(folder.name) / f"{safe_table_name(model_class)}.db"
+        if not dbpath.parent.exists():
+            raise RuntimeError(f'The sqlite storage folder "{dbpath.parent.absolute()}" was not found.')
+        log.debug(f"get sqlite3 connection at {dbpath}")
+        return sqlite3.connect(dbpath)
+
+    # NB using environment variables doesn't work
+    # monkeypatch.setenv("NZSHM22_HAZARD_STORE_LOCAL_CACHE", str(cache_folder.name))
+    monkeypatch.setattr(toshi_hazard_store.config, "LOCAL_CACHE_FOLDER", str(cache_folder))
+    monkeypatch.setattr(toshi_hazard_store.config, "SQLITE_ADAPTER_FOLDER", str(adapter_folder))
+    monkeypatch.setattr(
+        toshi_hazard_store.model.caching.cache_store,
+        "get_connection",
+        partial(temporary_cache_connection, folder=cache_folder),
+    )
+    monkeypatch.setattr(
+        toshi_hazard_store.db_adapter.sqlite.sqlite_adapter,
+        "get_connection",
+        partial(temporary_adapter_connection, folder=adapter_folder),
+    )
+    monkeypatch.setattr(toshi_hazard_store.model.caching.cache_store, "cache_enabled", lambda: True)
 
 
 @pytest.fixture(scope="function", autouse=True)
@@ -33,14 +83,6 @@ def pytest_generate_tests(metafunc):
         metafunc.parametrize("adapted_hazagg_model", ["pynamodb", "sqlite"], indirect=True)
     if "adapted_meta_model" in metafunc.fixturenames:
         metafunc.parametrize("adapted_meta_model", ["pynamodb", "sqlite"], indirect=True)
-
-
-@pytest.fixture()
-def setenvvar(tmp_path):
-    # ref https://adamj.eu/tech/2020/10/13/how-to-mock-environment-variables-with-pytest/
-    envvars = {"THS_SQLITE_FOLDER": str(tmp_path), "THS_USE_SQLITE_ADAPTER": "TRUE"}
-    with mock.patch.dict(os.environ, envvars, clear=True):
-        yield  # This is the magical bit which restore the environment after
 
 
 # @pytest.fixture(scope="function")